refactor: migrate on sliced logic for MFLoader
